requirements.txt
aiofiles==23.1.0
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==3.7.1
arxiv==1.4.8
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0
auth0-python==4.3.0
backoff==2.2.1
bidict==0.22.1
certifi==2023.7.22
cffi==1.15.1
chainlit==0.6.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.1
chromadb==0.4.2
click==8.1.6
coloredlogs==15.0.1
cryptography==42.0.5
dataclasses-json==0.5.13
Deprecated==1.2.14
exceptiongroup==1.1.2
fastapi==0.110.1
fastapi-socketio==0.0.10
feedparser==6.0.10
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.4.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.0.1
importlib-resources==6.0.0
Jinja2==3.1.2
langchain==0.1.14
langchainplus-sdk==0.0.20
langsmith==0.0.14
MarkupSafe==2.1.3
marshmallow==3.20.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nodeenv==1.8.0
numexpr==2.8.4
numpy==1.25.1
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-exporter-otlp-proto-http==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-proto==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
overrides==7.3.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
posthog==3.0.1
prisma==0.9.1
protobuf==4.23.4
pulsar-client==3.2.0
pycparser==2.21
pydantic==1.10.11
PyJWT==2.8.0
PyMuPDF==1.22.5
pypdf==4.1.0
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-graphql-client==0.4.3
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
reportlab==4.0.4
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.37.2
sympy==1.12
syncer==2.0.3
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.65.0
twilio==8.5.0
tzdata==2023.3
uptrace==1.18.0
urllib3==2.0.4
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
youtube-search==2.1.2
zipp==3.16.2
